Refine honeyhive-skills-strands tracks: isolated folders, no shared module.
no-honeyhive is pure Strands (agent.py only). honeyhive-integrated is self-contained with tracing, eval_harness.py, and a 6-case eval (code check + LLM anti-guessing judge, parallel workers with per-datapoint Agent instances).
