+++ honeyhive-skills-strands/honeyhive-integrated/.env.example
@@ -1,6 +1,7 @@
 OPENAI_API_KEY=
 OPENAI_MODEL=gpt-5-mini
+EVAL_MAX_WORKERS=4
 
 HH_API_KEY=
 HH_PROJECT=Strands
-HH_API_URL=
+HH_API_URL=https://api.dp1.us.honeyhive.ai
