+++ honeyhive-skills-strands/no-honeyhive/.env.example
@@ -1,11 +1,9 @@
-# Required for agent + eval
 OPENAI_API_KEY=
-# Optional — defaults to gpt-5-mini (agent + judge)
+# Optional — defaults to gpt-5-mini
 OPENAI_MODEL=gpt-5-mini
 
-# HoneyHive — needed from Step 2 onward
+# HoneyHive — add before Step 2 (instrument skill)
 # https://app.honeyhive.ai/settings/project/keys
-HH_API_KEY=
-HH_PROJECT=Strands
-# Dedicated / self-host only — omit to use the multi-tenant default
-HH_API_URL=
+# HH_API_KEY=
+# HH_PROJECT=Strands
+# HH_API_URL=https://api.dp1.us.honeyhive.ai
