+++ honeyhive-skills-strands/no-honeyhive/pyproject.toml
@@ -5,6 +5,4 @@ requires-python = ">=3.11"
 dependencies = [
     "strands-agents[openai]>=1.19.0",
     "python-dotenv>=1.0.0",
-    "honeyhive",
-    "openai>=1.0.0",
 ]
