+++ honeyhive-skills-strands/no-honeyhive/requirements.txt
@@ -1,4 +1,2 @@
 strands-agents[openai]>=1.19.0
 python-dotenv>=1.0.0
-honeyhive
-openai>=1.0.0
